requirements.txt
absl-py==1.0.0
albumentations==1.1.0
antlr4-python3-runtime==4.8
appdirs==1.4.4
black==21.4b2
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.8
click==8.0.3
cloudpickle==2.0.0
coloredlogs==15.0.1
ConfigArgParse==1.5.3
cycler==0.11.0
Cython==0.29.24
detectron2==0.6+cu102
fonttools==4.28.2
future==0.18.2
fvcore==0.1.5.post20211023
google-auth==2.3.3
google-auth-oauthlib==0.4.6
grpcio==1.42.0
humanfriendly==10.0
hydra-core==1.1.1
idna==3.3
imageio==2.13.3
importlib-metadata==4.8.2
iopath==0.1.9
joblib==1.1.0
kiwisolver==1.3.2
Markdown==3.3.6
matplotlib==3.5.0
mypy-extensions==0.4.3
networkx==2.6.3
numpy==1.21.4
oauthlib==3.1.1
omegaconf==2.1.1
opencv-python==4.5.4.60
opencv-python-headless==4.5.4.60
packaging==21.3
pandas==1.3.4
pathspec==0.9.0
Pillow==8.4.0
portalocker==2.3.2
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.3
pydot==1.4.2
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
PyWavelets==1.2.0
PyYAML==6.0
qudida==0.0.4
regex==2021.11.10
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
scikit-image==0.19.0
scikit-learn==1.0.1
scipy==1.7.3
seaborn==0.11.2
setuptools-scm==6.3.2
Shapely==1.8.0
simplejson==3.17.6
six==1.16.0
tabulate==0.8.9
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
termcolor==1.1.0
threadpoolctl==3.0.0
tifffile==2021.11.2
toml==0.10.2
tomli==1.2.2
torch==1.10.0
torchaudio==0.10.0
torchvision==0.11.1
tqdm==4.62.3
typing-extensions==4.0.1
urllib3==1.26.7
Werkzeug==2.0.2
yacs==0.1.8
zipp==3.6.0
